requirements.txt
certifi==2018.4.16
chardet==3.0.4
cookies==2.2.1
dateutils==0.6.6
Django==2.0.2
djangorestframework==3.7.7
idna==2.7
python-dateutil==2.7.3
python-dotenv==0.8.2
pytz==2018.5
requests==2.19.1
requests-mock==1.5.0
responses==0.9.0
six==1.11.0
urllib3==1.23
